adding updates to git, 2 tests failing, one is one of my new tests, but one is existing and still needs to be fixed TestAuthFailure

+++ projects/client/RabbitMQ.Client/src/client/impl/ISession.cs
@@ -48,7 +48,7 @@ namespace RabbitMQ.Client.Impl
         /// <summary>
         /// Gets the channel number.
         /// </summary>
-        int ChannelNumber { get; }
+        ushort ChannelNumber { get; }
 
         /// <summary>
         /// Gets the close reason.
